+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Next
 
+### 2.3.1
+
 - Introduce `lowPrecision` to tweak precision when computing average value. Thanks @BenjaminVanRyseghem
 - Fix #411: spaceSeparated is not working for bytes. Thanks @BenjaminVanRyseghem